Moved stuff to assembly class; should be able to do send query and receive json response.
